pyproject.toml
[tool.poetry]
name = "tgisper"
version = "0.2.0"
description = "Telegram bot for recognizing voice messages based on Whisper models"
license = "MIT"
authors = [
    "ckaytev",
]
readme = "README.md"
repository = "https://github.com/ckaytev/tgisper"

[tool.poetry.dependencies]
python = ">=3.8.1, <3.9"
pytelegrambotapi = "^4.9.0"
numpy = "^1.24.3"
faster-whisper = "0.5.1"
ffmpeg-python = "^0.2.0"

[tool.poetry.scripts]
tgisper = "tgisper.tgisperbot:main"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
